+++ src/main/sh/app/commands/doBackupDevOPS.sh
@@ -1,0 +1,8 @@
+_doBackupDevOPS() {
+    currentBackupDir=$currentBackups/devops
+
+    mkdir -p $currentBackupDir/
+
+    cp $DIR_DEVOPS/.env $currentBackupDir/docker-compose.env.backup
+    cp $DIR_DEVOPS/docker-compose.yml $currentBackupDir/docker-compose.yml.backup
+}