+++ src/main/sh/app/commands/doBackupDockerRegistry.sh
@@ -1,0 +1,7 @@
+_doBackupDockerRegistry() {
+    currentBackupDir=$currentBackups/registry
+
+    mkdir -p $currentBackupDir/
+
+    ## TODO: backup registry
+}