+++ src/main/sh/app/commands/doBackupMongoDB.sh
@@ -1,0 +1,7 @@
+_doBackupMongoDB() {
+    currentBackupDir=$currentBackups/db/mongodb
+
+    mkdir -p $currentBackupDir/
+
+    ## TODO: backup mongodb
+}