+++ src/main/sh/app/commands/doBackupPostgres.sh
@@ -1,0 +1,7 @@
+_doBackupPostgres() {
+    currentBackupDir=$currentBackups/db/postgres
+
+    mkdir -p $currentBackupDir/
+
+    ## TODO: backup postgres
+}