+++ src/main/sh/app/commands/doDockerClean.sh
@@ -1,0 +1,7 @@
+_doDockerClean() {
+    docker container kill --force $(docker container ls -aq)
+    docker container rm --force $(docker container ls -aq)
+    docker volume rm --force $(docker volume ls -aq)
+    docker image rm --force $(docker image ls -aq)
+    docker network rm --force $(docker network ls -aq)
+}