+++ src/main/sh/app/tasks/taskInstall.sh
@@ -1,0 +1,17 @@
+install() {
+    _doInstallTools
+
+    _doInstallDocker
+    _doInstallDockerCompose
+
+    _doInstallDevOps
+
+    for mode in ${CRON_BACKUP_MODES//,/ }; do
+        _doInstallCron "$mode"
+    done
+
+    _doDockerComposeUp
+
+    _doGarbageCollect
+    _doTerminalClean
+}