+++ src/main/sh/utils/remove-folder.sh
@@ -1,0 +1,3 @@
+if [ -d "$1" ]; then
+    rm -rf $1/
+fi