+++ CHANGELOG.md
@@ -3,7 +3,13 @@
 All notable changes to this project will be documented in this file. See [standard-version](https://github.com/conventional-changelog/standard-version) for commit guidelines.
 
 
-## [0.2.0](https://github.com/lildutils/ldu-devops/releases/tag/0.2.0) (2023-03-28)
+## [0.3.0](https://github.com/lildutils/ldu-devops-docker/releases/tag/0.3.0) (2023-03-29)
+
+- FEAT: Test Framework next steps
+- FEAT: Application compiler for make.sh script file
+- FEAT: Nginx docker compose service
+
+## [0.2.0](https://github.com/lildutils/ldu-devops-docker/releases/tag/0.2.0) (2023-03-28)
 
 - FEAT: Autorun for make.sh
 - FEAT: Wait for check logs and than clear terminal
@@ -14,7 +20,7 @@ All notable changes to this project will be documented in this file. See [standa
 - FEAT: Fail2Ban max retry dynamic setting
 - FEAT: Gitlab SMTP enabled or not
 
-## [0.1.0](https://github.com/lildutils/ldu-devops/releases/tag/0.1.0) (2023-03-28)
+## [0.1.0](https://github.com/lildutils/ldu-devops-docker/releases/tag/0.1.0) (2023-03-28)
 
 - FEAT: Build process with clean option
 - FEAT: README markdown file created
@@ -24,7 +30,7 @@ All notable changes to this project will be documented in this file. See [standa
 - FEAT: Docker resources for make devops
 - FEAT: Make devops script created
 
-## [0.0.0](https://github.com/lildutils/ldu-devops/releases/tag/0.0.0) (2023-03-28)
+## [0.0.0](https://github.com/lildutils/ldu-devops-docker/releases/tag/0.0.0) (2023-03-28)
 
 - FEAT: Project file created
 - FEAT: Initial commit